requirements.txt
# Core scientific stack (used throughout code/ and notebooks/)
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
scipy>=1.10

# Parquet engine for pandas.to_parquet / read_parquet
pyarrow>=14.0

# ML + statistics (used in code/ and referenced in the book)
scikit-learn>=1.3
statsmodels>=0.14
torch>=2.1

# Time-series and plotting helpers used in code/
mplfinance>=0.12
plotly>=5.0

# HTTP and API clients for external data wrappers
requests>=2.31
v20>=3.0
trading_ig>=0.0.22

# Streaming and lightweight logging helpers (used in code/streaming)
pyzmq>=25.0
q>=2.7

# Interactive + notebook tooling (used by tools/validate_notebooks.py and notebooks/)
ipython>=8.0
jupyterlab>=4.0
ipykernel>=6.0
nbformat>=5.9
nbclient>=0.10

# LaTeX minted (provides the `pygmentize` executable used by minted)
Pygments>=2.16
